Refactor specialize
