+++ core/src/parser/op.rs
@@ -241,3 +241,40 @@ impl Operator {
         }
     }
 }
+
+pub(super) const COMMA_SYM: &str = ",";
+pub(super) const ASSIGNMENT_SYM: &str = "=";
+pub(super) const MEMBER_ACCESS_SYM: &str = ".";
+
+/// Default operators used in the language grammar and required for parsing. \
+/// These operators are defined in the parser and are required to produce valid ASTs. \
+/// The binary operators are replaced with `Ast` nodes by `syntax_sugar::specialize` after parsing a `parse_top_expr` expression.
+/// - `semicolon`: `;` - Used to separate statements becomes an `Ast::Block` node.
+/// - `comma`: `,` - Used to separate expressions in tuples and lists becomes an `Ast::Tuple` or `Ast::List` node.
+/// - `assignment`: `=` - Used to assign values to variables becomes an `Ast::Assignment` node.
+/// - `member access`: `.` - Used to access members of records becomes an `Ast::MemberAccess` node.
+pub fn default_operators() -> Vec<OpInfo> {
+    vec![
+        OpInfo {
+            symbol: COMMA_SYM.to_string(),
+            position: OpPos::Infix,
+            precedence: prec::COMMA,
+            associativity: OpAssoc::Left,
+            allow_trailing: true,
+        },
+        OpInfo {
+            symbol: ASSIGNMENT_SYM.to_string(),
+            position: OpPos::Infix,
+            precedence: prec::ASSIGNMENT,
+            associativity: OpAssoc::Right,
+            allow_trailing: false,
+        },
+        OpInfo {
+            symbol: MEMBER_ACCESS_SYM.to_string(),
+            position: OpPos::Infix,
+            precedence: prec::MEMBER_ACCESS,
+            associativity: OpAssoc::Left,
+            allow_trailing: false,
+        },
+    ]
+}
